Move RemoteAST's RemoteASTTypeBuilder to lib/AST/ASTDemangler.cpp
