+++ lib/AST/CMakeLists.txt
@@ -7,6 +7,7 @@ add_swift_host_library(swiftAST STATIC
   AccessScopeChecker.cpp
   AccessRequests.cpp
   ASTContext.cpp
+  ASTDemangler.cpp
   ASTDumper.cpp
   ASTMangler.cpp
   ASTNode.cpp
